requirements.txt
# 深度学习框架
torch>=1.10.0
transformers>=4.20.0

# CRF层
pytorch-crf>=0.7.2

# 数据处理
numpy>=1.20.0
pandas>=1.3.0

# NER评估
seqeval>=1.2.2

# 机器学习工具
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
spacy>=3.0.0

# 配置管理
pyyaml>=5.4.0

# 进度条
tqdm>=4.60.0

# Jupyter支持
jupyter>=1.0.0
ipython>=7.30.0

# 工具库
requests>=2.26.0
